# docs/swagger.yaml
basePath: /
definitions:
  internal_app_http_handler.ErrorResponse:
    properties:
      error:
        example: email already registered
        type: string
    type: object
  internal_app_http_handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securePass1
        type: string
    required:
    - email
    - password
    type: object
  internal_app_http_handler.RefreshRequest:
    properties:
      refresh_token:
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4
        type: string
    required:
    - refresh_token
    type: object
  internal_app_http_handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: securePass1
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_app_http_handler.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4
        type: string
    type: object
  internal_app_http_handler.UserResponse:
    properties:
      created_at:
        example: "2024-01-15T10:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Ivan Ivanov
        type: string
      role:
        example: buyer
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: george200135@gmail.com
    name: leenwood
  description: Authentication and authorization service for the marketplace.
  title: Auth Service API
  version: "1.0"
paths:
  /.well-known/jwks.json:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: JWKS — RSA public key for JWT verification
      tags:
      - system
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_http_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_http_handler.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      summary: Authenticate user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_http_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      summary: Invalidate refresh token
      tags:
      - auth
  /api/v1/auth/me:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete current user account
      tags:
      - auth
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_http_handler.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_http_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_http_handler.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_http_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_app_http_handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Liveness probe
      tags:
      - system
  /ready:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_app_http_handler.ErrorResponse'
      summary: Readiness probe — checks Postgres and Redis
      tags:
      - system
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Paste your JWT access token: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"